+++ test/integration/loans.test.ts
@@ -176,357 +176,6 @@ describe('Loans', () => {
               const loanStatus = helpers.details.loan.status
               expect(loanStatus).to.equal(2)
             })
-
-            it('should be able to create and repay a loan', async () => {
-              const borrower = await getNamedSigner('borrower')
-
-              const { nfts, getHelpers } = await takeOutLoanWithNfts(hre, {
-                amount: 100,
-                lendToken: market.lendingToken,
-                borrower,
-                version: 1,
-              })
-
-              const helpers: LoanHelpersReturn = await getHelpers()
-
-              expect(helpers.details.loan).to.exist
-
-              console.log('helpers.details.loan', helpers.details.loan)
-
-              const loanId = helpers.details.loan.id
-              console.log('loanId', loanId)
-
-              //  const lHelpers = await loanHelpers(loanId)
-
-              //need to give fake ERC20 to borrower
-              //and add fake ERC20 to the borrow escrow
-
-              const lendingToken =
-                typeof market.lendingToken === 'string'
-                  ? await tokens.get(market.lendingToken)
-                  : market.lendingToken
-
-              const lendingTokenDecimals = await lendingToken.decimals()
-              console.log('decimals', lendingTokenDecimals)
-
-              await getFunds({
-                to: borrower,
-                tokenSym: market.lendingToken,
-                amount: 200 * 10 ** lendingTokenDecimals,
-                hre,
-              })
-
-              const borrowerAddress = await borrower.getAddress()
-              const borrowerBalance = await lendingToken.balanceOf(
-                borrowerAddress
-              )
-
-              console.log('balanceOf', borrowerBalance.toString())
-
-              console.log('market.lendingToken', market.lendingToken)
-
-              const balanceLeftToRepay = await diamond.getTotalOwed(loanId)
-              console.log('balanceLeftToRepay', balanceLeftToRepay)
-
-              await lendingToken
-                .connect(ethers.provider.getSigner(borrowerAddress))
-                .approve(diamond.address, balanceLeftToRepay)
-
-              //let response = await helpers.repay( 100000000 , borrower )
-
-              await diamond.connect(borrower).repayLoan(loanId, 100010000)
-
-              const totalOwedAfterRepay = await diamond.getTotalOwed(loanId)
-
-              console.log('totalOwedAfterRepay', totalOwedAfterRepay.toString())
-
-              const loanData = await diamond.getLoan(loanId)
-              console.log('loanData')
-
-              console.log('balanceLeftToRepay 2', balanceLeftToRepay)
-
-              console.log('balanceOf After', borrowerBalance.toString())
-
-              expect(parseInt(borrowerBalance.toString())).to.equal(200000000)
-
-              expect(loanData.status).to.equal(3) //3 = repaid
-
-              expect(parseInt(totalOwedAfterRepay.toString())).to.equal(0)
-            })
-
-            it('should be able to create and escrow-repay a loan', async () => {
-              const borrower = await getNamedSigner('borrower')
-
-              const { nfts, getHelpers } = await takeOutLoanWithNfts(hre, {
-                amount: 100,
-                lendToken: market.lendingToken,
-                borrower,
-                version: 1,
-              })
-
-              const helpers: LoanHelpersReturn = await getHelpers()
-
-              expect(helpers.details.loan).to.exist
-
-              console.log('helpers.details.loan', helpers.details.loan)
-
-              const loanId = helpers.details.loan.id
-              console.log('loanId', loanId)
-
-              //  const lHelpers = await loanHelpers(loanId)
-
-              //need to give fake ERC20 to borrower
-              //and add fake ERC20 to the borrow escrow
-
-              const lendingToken =
-                typeof market.lendingToken === 'string'
-                  ? await tokens.get(market.lendingToken)
-                  : market.lendingToken
-
-              const lendingTokenDecimals = await lendingToken.decimals()
-              console.log('decimals', lendingTokenDecimals)
-
-              await getFunds({
-                to: borrower,
-                tokenSym: market.lendingToken,
-                amount: 200 * 10 ** lendingTokenDecimals,
-                hre,
-              })
-
-              const borrowerAddress = await borrower.getAddress()
-              const borrowerBalance = await lendingToken.balanceOf(
-                borrowerAddress
-              )
-
-              console.log('balanceOf', borrowerBalance.toString())
-
-              console.log('market.lendingToken', market.lendingToken)
-
-              const balanceLeftToRepay = await diamond.getTotalOwed(loanId)
-              console.log('balanceLeftToRepay', balanceLeftToRepay.toString())
-
-              await lendingToken
-                .connect(ethers.provider.getSigner(borrowerAddress))
-                .approve(diamond.address, balanceLeftToRepay)
-
-              //let response = await helpers.repay( 100000000 , borrower )
-
-              await diamond.connect(borrower).escrowRepay(loanId, 100010000)
-
-              const totalOwedAfterRepay = await diamond.getTotalOwed(loanId)
-
-              console.log('totalOwedAfterRepay', totalOwedAfterRepay.toString())
-
-              const loanData = await diamond.getLoan(loanId)
-              console.log('loanData')
-
-              console.log('balanceOf After', borrowerBalance.toString())
-
-              console.log('balanceLeftToRepay 2', balanceLeftToRepay.toString())
-
-              expect(loanData.status).to.equal(3) //3 = repaid
-
-              expect(parseInt(borrowerBalance.toString())).to.equal(299990000)
-
-              expect(parseInt(totalOwedAfterRepay.toString())).to.equal(0)
-            })
-
-            it('should be able to create, earn profits on, and repay a loan', async () => {
-              const borrower = await getNamedSigner('borrower')
-
-              const { nfts, getHelpers } = await takeOutLoanWithNfts(hre, {
-                amount: 100,
-                lendToken: market.lendingToken,
-                borrower,
-                version: 1,
-              })
-
-              const helpers: LoanHelpersReturn = await getHelpers()
-
-              expect(helpers.details.loan).to.exist
-
-              console.log('helpers.details.loan', helpers.details.loan)
-
-              const loanId = helpers.details.loan.id
-              console.log('loanId', loanId)
-
-              //  const lHelpers = await loanHelpers(loanId)
-
-              //need to give fake ERC20 to borrower
-              //and add fake ERC20 to the borrow escrow
-
-              const lendingToken =
-                typeof market.lendingToken === 'string'
-                  ? await tokens.get(market.lendingToken)
-                  : market.lendingToken
-
-              const lendingTokenDecimals = await lendingToken.decimals()
-              console.log('decimals', lendingTokenDecimals)
-
-              await getFunds({
-                to: borrower,
-                tokenSym: market.lendingToken,
-                amount: 200 * 10 ** lendingTokenDecimals,
-                hre,
-              })
-
-              const borrowerAddress = await borrower.getAddress()
-              const borrowerBalance = await lendingToken.balanceOf(
-                borrowerAddress
-              )
-
-              console.log('balanceOf', borrowerBalance.toString())
-
-              console.log('market.lendingToken', market.lendingToken)
-
-              const balanceLeftToRepay = await diamond.getTotalOwed(loanId)
-              console.log('balanceLeftToRepay', balanceLeftToRepay.toString())
-
-              await lendingToken
-                .connect(ethers.provider.getSigner(borrowerAddress))
-                .approve(diamond.address, balanceLeftToRepay)
-
-              await diamond.connect(borrower).repayLoan(loanId, 100010000)
-
-              const totalOwedAfterRepay = await diamond.getTotalOwed(loanId)
-
-              console.log('totalOwedAfterRepay', totalOwedAfterRepay.toString())
-
-              const loanData = await diamond.getLoan(loanId)
-              console.log('loanData')
-
-              console.log('balanceLeftToRepay 2', balanceLeftToRepay.toString())
-
-              console.log('balanceOf After', borrowerBalance.toString())
-
-              expect(parseInt(borrowerBalance.toString())).to.equal(499980000)
-
-              expect(loanData.status).to.equal(3) //3 = repaid
-
-              expect(parseInt(totalOwedAfterRepay.toString())).to.equal(0)
-            })
-
-            it('should be able to create, earn profits on, and repay a loan with excess profit ', async () => {
-              const borrower = await getNamedSigner('borrower')
-
-              const { nfts, getHelpers } = await takeOutLoanWithNfts(hre, {
-                amount: 100,
-                lendToken: market.lendingToken,
-                borrower,
-                version: 1,
-              })
-
-              const helpers: LoanHelpersReturn = await getHelpers()
-
-              expect(helpers.details.loan).to.exist
-
-              console.log('helpers.details.loan', helpers.details.loan)
-
-              // console.log('helpers.details.loan', helpers.details.loan)
-
-              const loanId = helpers.details.loan.id
-              console.log('loanId', loanId)
-
-              //  const lHelpers = await loanHelpers(loanId)
-
-              //need to give fake ERC20 to borrower
-              //and add fake ERC20 to the borrow escrow
-
-              const lendingToken =
-                typeof market.lendingToken === 'string'
-                  ? await tokens.get(market.lendingToken)
-                  : market.lendingToken
-
-              const lendingTokenDecimals = await lendingToken.decimals()
-              console.log('decimals', lendingTokenDecimals)
-
-              await getFunds({
-                to: borrower,
-                tokenSym: market.lendingToken,
-                amount: 200 * 10 ** lendingTokenDecimals,
-                hre,
-              })
-
-              //simulate that excess profit was made by the user on the borrowed money
-              const borrowedFundsEscrowAddress = await diamond
-                .connect(borrower)
-                .getLoanEscrow(loanId)
-
-              await getFunds({
-                to: borrowedFundsEscrowAddress,
-                tokenSym: market.lendingToken,
-                amount: 30 * 10 ** lendingTokenDecimals,
-                hre,
-              })
-
-              const tTokenAddress = await diamond
-                .connect(borrower)
-                .getTTokenFor(lendingToken.address)
-
-              const originalTTokenEscrowBalanceBN =
-                await lendingToken.balanceOf(tTokenAddress)
-
-              const originalTTokenEscrowBalance: number = parseInt(
-                originalTTokenEscrowBalanceBN.toString()
-              )
-
-              expect(originalTTokenEscrowBalance).to.equal(99700030000)
-
-              let loanEscrowBalance = await lendingToken.balanceOf(
-                borrowedFundsEscrowAddress
-              )
-
-              expect(loanEscrowBalance).to.equal(130000000)
-
-              const borrowerAddress = await borrower.getAddress()
-              const borrowerBalance = await lendingToken.balanceOf(
-                borrowerAddress
-              )
-
-              console.log('balanceOf', borrowerBalance.toString())
-
-              console.log('market.lendingToken', market.lendingToken)
-
-              const balanceLeftToRepay = await diamond.getTotalOwed(loanId)
-              console.log('balanceLeftToRepay', balanceLeftToRepay.toString())
-
-              await lendingToken
-                .connect(ethers.provider.getSigner(borrowerAddress))
-                .approve(diamond.address, balanceLeftToRepay)
-
-              await diamond.connect(borrower).repayLoan(loanId, 100010000)
-
-              const totalOwedAfterRepay = await diamond.getTotalOwed(loanId)
-
-              const loanData = await diamond.getLoan(loanId)
-
-              const amtRepaid = 100010000
-              const expectedProfitFee = 1500000
-
-              //The difference should be the loan repayment + profit fee
-              const finalTTokenEscrowBalance = await lendingToken.balanceOf(
-                tTokenAddress
-              )
-              console.log(
-                'finalTTokenEscrowBalance 2 ',
-                finalTTokenEscrowBalance.toString()
-              )
-              expect(parseInt(finalTTokenEscrowBalance.toString())).to.equal(
-                originalTTokenEscrowBalance + expectedProfitFee + amtRepaid
-              ) //99700030000
-
-              loanEscrowBalance = await lendingToken.balanceOf(
-                borrowedFundsEscrowAddress
-              )
-              console.log('loanEscrowBalance 3 ', loanEscrowBalance.toString())
-              expect(loanEscrowBalance).to.equal(0)
-
-              expect(parseInt(borrowerBalance.toString())).to.equal(599970000)
-
-              expect(loanData.status).to.equal(3) //3 = repaid
-
-              expect(parseInt(totalOwedAfterRepay.toString())).to.equal(0)
-            })
           })
         }
